convert NPCIndex to uint32

+++ packages/contracts/src/libraries/LibNPC.sol
@@ -21,7 +21,7 @@ library LibNPC {
   function create(
     IWorld world,
     IUintComp components,
-    uint256 index,
+    uint32 index,
     string memory name,
     uint256 roomIndex
   ) internal returns (uint256) {
@@ -62,7 +62,7 @@ library LibNPC {
   /////////////////
   // GETTERS
 
-  function getIndex(IUintComp components, uint256 id) internal view returns (uint256) {
+  function getIndex(IUintComp components, uint256 id) internal view returns (uint32) {
     return IndexNPCComponent(getAddressById(components, IndexNPCCompID)).getValue(id);
   }
 
@@ -79,7 +79,7 @@ library LibNPC {
   // QUERIES
 
   // Return the ID of a Merchant by its index
-  function getByIndex(IUintComp components, uint256 index) internal view returns (uint256 result) {
+  function getByIndex(IUintComp components, uint32 index) internal view returns (uint256 result) {
     QueryFragment[] memory fragments = new QueryFragment[](2);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsNPCCompID), "");
     fragments[1] = QueryFragment(
